[DEV 2.5] Integrated the Undo feature and Command Pattern to work with the Search Button. Added attributes to the Country Class to make this possible.
