+++ src/CommandOperation/CommandHistory.java
@@ -1,13 +1,16 @@
 package src.CommandOperation;
 
+import src.Country;
+
 import java.util.ArrayList;
+import java.util.HashMap;
 
 /**
  * The CommandHistory class acts as the Iterable class to store
  * commands.
  */
 public class CommandHistory {
-    private ArrayList<Command> Previous;
+    public ArrayList<Command> Previous;
 
     public CommandHistory() {
         Previous = new ArrayList<Command>();
@@ -56,5 +59,13 @@ public class CommandHistory {
         for (Command c: Previous){
             System.out.println(c);
         }
+    }
+
+    public HashMap<String, Command> getCommandHashMap () {
+        HashMap<String, Command> result = new HashMap<>();
+        for (Command c: Previous){
+            result.put(c.getInfo().getName(), c);
+        }
+        return result;
     }
 }
