+++ src/CountryTest.java
@@ -1,0 +1,17 @@
+package src;
+
+import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.*;
+
+/**
+ * The Country class to test the Country object for Data-Visualisation purposes.
+ * Tests various components of the class.
+ */
+public class CountryTest {
+
+    @Test
+    void ContinentTest() {
+        Country Canada = new Country("Canada", 1.5);
+        assertEquals("Americas", Canada.Continent);
+    }
+}
