+++ src/SearchButton.java
@@ -1,3 +1,5 @@
+package src;
+
 import javafx.scene.control.Button;
 import javafx.scene.control.TextField;
 import javafx.scene.input.MouseEvent;
